Rename Torus-direct to CustomAuth
