+++ RNCustomAuthSdk.podspec
@@ -21,6 +21,6 @@ Pod::Spec.new do |s|
 
   s.dependency 'React'
   s.dependency 'secp256k1.c', '0.1.2'
-  s.dependency 'Torus-directSDK', '~> 0.2.0'
+  s.dependency 'CustomAuth', '~> 2.0.0'
   #s.dependency "others"
 end
