+++ android/build.gradle
@@ -49,6 +49,6 @@ repositories {
 
 dependencies {
     api 'com.facebook.react:react-native:+'
-    api 'org.torusresearch:customauth-android-sdk:2.0.0'
+    api 'org.torusresearch:customauth-android-sdk:2.0.4'
 }
