+++ android/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="org.torusresearch.rntorusdirect">
+          package="org.torusresearch.rncustomauth">
 
 </manifest>
