+++ example/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "org.torusresearch.torusdirectexample",
+    package = "org.torusresearch.customauthexample",
 )
 
 android_resource(
     name = "res",
-    package = "org.torusresearch.torusdirectexample",
+    package = "org.torusresearch.customauthexample",
     res = "src/main/res",
 )
