+++ example/android/app/src/main/java/org/torusresearch/torusdirectexample/MainActivity.java
@@ -1,4 +1,4 @@
-package org.torusresearch.torusdirectexample;
+package org.torusresearch.customauthexample;
 
 import android.os.Bundle;
