+++ example/android/app/src/main/java/org/torusresearch/torusdirectexample/MainApplication.java
@@ -1,4 +1,4 @@
-package org.torusresearch.torusdirectexample;
+package org.torusresearch.customauthexample;
 
 import android.app.Application;
 import android.content.Context;
@@ -11,7 +11,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import org.torusresearch.torusdirectexample.generated.BasePackageList;
+import org.torusresearch.customauthexample.generated.BasePackageList;
 
 import org.unimodules.adapters.react.ReactAdapterPackage;
 import org.unimodules.adapters.react.ModuleRegistryAdapter;
