+++ example2/ios/Podfile
@@ -4,8 +4,9 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 platform :ios, '11.0'
 
 target 'example2' do
-  config = use_native_modules!
+  use_modular_headers!
 
+  config = use_native_modules!
 
   pod 'secp256k1.c', '0.1.2', :modular_headers => true
