+++ ios/RNCustomAuthSdk.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:RNTorusDirectSDK.xcodeproj">
+      location = "group:RNCustomAuthSDK.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
